phone_dictonary v0.1.0 PhoneDictonary

PhoneDictonary module for mapping phone number to given dictonary of words

Functions

add_prefix(list, prefix, acc)

this method will combine input existing prefix whith the current chars and if accumalator setup

zero then it will match pattern and return same list input

combine(n, list, acc)

this method will combine given number to correspond mapping recursively

getWords(len, str, word_groups, dict_map)

method for getting words from dictonary recursively

used String.split_at which is faster than String.slice

get_sub_str(str, dict_map)

method for creating sets of substring of words

number_to_words(n)

this method will take number of ten digit as a input

to_map(list, acc)

#this method will convert list of string to map recursively